+++ account_move_kmitl/i18n/th.po
@@ -1,0 +1,105 @@
+# Translation of Odoo Server.
+# This file contains the translation of the following modules:
+# 	* account_move_kmitl
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: Odoo Server 16.0\n"
+"Report-Msgid-Bugs-To: \n"
+"POT-Creation-Date: 2026-03-11 00:00+0000\n"
+"PO-Revision-Date: 2026-03-11 07:00+0700\n"
+"Last-Translator: \n"
+"Language-Team: \n"
+"Language: th\n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: 8bit\n"
+"Plural-Forms: nplurals=1; plural=0;\n"
+
+#. module: account_move_kmitl
+#: model:ir.model,name:account_move_kmitl.model_account_move
+msgid "Journal Entry"
+msgstr "รายการบันทึกบัญชี"
+
+#. module: account_move_kmitl
+#: model:ir.model.fields,field_description:account_move_kmitl.field_account_move__state
+msgid "Status"
+msgstr "สถานะ"
+
+#. module: account_move_kmitl
+#: model:ir.model.fields.selection,name:account_move_kmitl.selection__account_move__state__submitted
+msgid "Submitted"
+msgstr "ส่งอนุมัติ"
+
+#. module: account_move_kmitl
+#: model:ir.model.fields,field_description:account_move_kmitl.field_account_move__budget_commitment_id
+msgid "Budget Commitment"
+msgstr "ผูกพันงบประมาณ"
+
+#. module: account_move_kmitl
+#: model:ir.model.fields,field_description:account_move_kmitl.field_account_move__budget_account_id
+msgid "Budget Account"
+msgstr "บัญชีงบประมาณ"
+
+#. module: account_move_kmitl
+#: model:ir.model.fields,help:account_move_kmitl.field_account_move__budget_commitment_id
+msgid "Related budget commitment for this journal entry"
+msgstr "ผูกพันงบประมาณที่เกี่ยวข้องกับรายการบันทึกบัญชีนี้"
+
+#. module: account_move_kmitl
+#: model:ir.model.fields,help:account_move_kmitl.field_account_move__budget_account_id
+msgid "Budget account to be used for commitment"
+msgstr "บัญชีงบประมาณที่ใช้สำหรับการผูกพัน"
+
+#. module: account_move_kmitl
+#: model_terms:ir.ui.view,arch_db:account_move_kmitl.view_move_form_inherit_kmitl
+msgid "Submit"
+msgstr "ส่งอนุมัติ"
+
+#. module: account_move_kmitl
+#: model_terms:ir.ui.view,arch_db:account_move_kmitl.view_move_form_inherit_kmitl
+msgid "Reset to Draft"
+msgstr "ตั้งค่าเป็นฉบับร่าง"
+
+#. module: account_move_kmitl
+#: model_terms:ir.ui.view,arch_db:account_move_kmitl.view_move_form_inherit_kmitl
+msgid "Budget"
+msgstr "งบประมาณ"
+
+#. module: account_move_kmitl
+#: model_terms:ir.ui.view,arch_db:account_move_kmitl.view_move_form_inherit_kmitl
+msgid "View"
+msgstr "ดู"
+
+#. module: account_move_kmitl
+#: model_terms:ir.ui.view,arch_db:account_move_kmitl.view_move_form_inherit_kmitl
+msgid "Budget Commitment"
+msgstr "ผูกพันงบประมาณ"
+
+#. module: account_move_kmitl
+#: code:addons/account_move_kmitl/models/account_move.py:0
+#, python-format
+msgid "Only draft entries can be submitted."
+msgstr "เฉพาะรายการฉบับร่างเท่านั้นที่สามารถส่งอนุมัติได้"
+
+#. module: account_move_kmitl
+#: code:addons/account_move_kmitl/models/account_move.py:0
+#, python-format
+msgid "Only submitted entries can be reset to draft."
+msgstr "เฉพาะรายการที่ส่งอนุมัติแล้วเท่านั้นที่สามารถตั้งค่าเป็นฉบับร่างได้"
+
+#. module: account_move_kmitl
+#: code:addons/account_move_kmitl/models/account_move.py:0
+#, python-format
+msgid "No budget commitment linked to this journal entry"
+msgstr "ไม่มีผูกพันงบประมาณที่เชื่อมโยงกับรายการบันทึกบัญชีนี้"
+
+#. module: account_move_kmitl
+#: model:tier.definition,name:account_move_kmitl.tier_definition_account_move_head_accountant
+msgid "หัวหน้าฝ่ายบัญชี"
+msgstr ""
+
+#. module: account_move_kmitl
+#: model:tier.definition,name:account_move_kmitl.tier_definition_account_move_accountant
+msgid "ผู้จัดทำ/ผู้ตรวจสอบ"
+msgstr ""
