+++ account_move_kmitl/views/account_move_views.xml
@@ -1,23 +1,48 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
 
-    <!-- Account Move Form View with Budget Integration -->
-    <record id="budget_view_move_form" model="ir.ui.view">
-        <field name="name">account.move.form.budget</field>
+    <record id="view_move_form_inherit_kmitl" model="ir.ui.view">
+        <field name="name">account.move.form.inherit.kmitl</field>
         <field name="model">account.move</field>
         <field name="inherit_id" ref="account.view_move_form"/>
         <field name="arch" type="xml">
 
-            <!-- Add Budget tab -->
+            <!-- Statusbar: draft -> submitted -> posted -->
+            <xpath expr="//field[@name='state']" position="attributes">
+                <attribute name="statusbar_visible">draft,submitted,posted</attribute>
+            </xpath>
+
+            <!-- Hidden fields for tier validation -->
+            <xpath expr="//field[@name='id']" position="after">
+                <field name="validation_status" invisible="1"/>
+                <field name="need_validation" invisible="1"/>
+            </xpath>
+
+            <!-- Submit and Reset to Draft buttons -->
+            <xpath expr="//button[@name='action_post'][2]" position="after">
+                <button
+                    name="action_submit"
+                    string="Submit"
+                    type="object"
+                    class="oe_highlight"
+                    attrs="{'invisible': [('state', '!=', 'draft')]}"
+                />
+                <button
+                    name="action_draft"
+                    string="Reset to Draft"
+                    type="object"
+                    attrs="{'invisible': ['|', ('state', '!=', 'submitted'), ('validation_status', 'in', ['validated', 'rejected', 'pending'])]}"
+                />
+            </xpath>
+
+            <!-- Budget tab -->
             <xpath expr="//notebook" position="inside">
                 <page string="Budget" name="budget">
-
                     <group string="Budget">
                         <group name="budget_info">
-                            <field name="budget_commitment_id" attrs="{'readonly': [('state', '=', 'posted')]}" />
+                            <field name="budget_commitment_id" attrs="{'readonly': [('state', '=', 'posted')]}"/>
                             <field name="analytic_distribution" widget="analytic_distribution" readonly="1"/>
                         </group>
-
                         <group name="analytic_dimensions">
                             <field name="activity_analytic_id" domain="[('plan_id.code', '=', 'activities')]" options="{'no_create': True}" attrs="{'readonly': [('state', '=', 'posted')]}"/>
                             <field name="department_analytic_id" domain="[('plan_id.code', '=', 'departments')]" options="{'no_create': True}" attrs="{'readonly': [('state', '=', 'posted')]}"/>
@@ -26,7 +51,6 @@
                             <field name="budget_account_id" options="{'no_create': True}" attrs="{'readonly': [('state', '=', 'posted')]}"/>
                         </group>
                     </group>
-
                     <div class="oe_button_box" name="button_box_budget">
                         <button name="action_view_budget_commitment" type="object" class="oe_stat_button" icon="fa-calculator" attrs="{'invisible': [('budget_commitment_id', '=', False)]}">
                             <div class="o_field_widget o_stat_info">
@@ -35,7 +59,6 @@
                             </div>
                         </button>
                     </div>
-
                 </page>
             </xpath>
