--- account_move_submitted_state/__manifest__.py
@@ -1,18 +1,0 @@
-{
-    "name": "Account Move Submitted State",
-    "version": "16.0.1.0.0",
-    "category": "Accounting",
-    "summary": "Add Submitted State to Account Moves for Approval Workflow",
-    "author": "KMITL",
-    "website": "https://www.kmitl.ac.th",
-    "license": "AGPL-3",
-    "depends": [
-        "account",
-    ],
-    "data": [
-        "views/account_move_views.xml",
-    ],
-    "installable": True,
-    "auto_install": False,
-    "application": False,
-}
