--- account_move_submitted_state/views/account_move_views.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <!-- Inherit account.move form view to add submitted state statusbar and button -->
-    <record id="view_move_form_inherit_submitted_state" model="ir.ui.view">
-        <field name="name">account.move.form.inherit.submitted.state</field>
-        <field name="model">account.move</field>
-        <field name="inherit_id" ref="account.view_move_form"/>
-        <field name="arch" type="xml">
-
-            <xpath expr="//field[@name='state']" position="attributes">
-                <attribute name="statusbar_visible">draft,submitted,posted</attribute>
-            </xpath>
-
-            <xpath expr="//button[@name='action_post'][2]" position="after">
-                <button name="action_submit" string="Submit" type="object" class="oe_highlight" attrs="{'invisible': [('state', '!=', 'draft')]}"/>
-                <button name="action_draft" string="Reset to Draft" type="object" attrs="{'invisible': [('state', '!=', 'submitted')]}"/>
-            </xpath>
-
-        </field>
-    </record>
-</odoo>
