--- account_move_tier_validation_kmitl/models/__init__.py
@@ -1,3 +1,0 @@
-# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
-
-from . import account_move
