--- account_move_tier_validation_kmitl/views/account_move_views.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<odoo>
-    <!-- Inherit from account_move_submitted_state view to integrate tier validation -->
-    <record id="view_move_form_inherit_tier_validation_kmitl" model="ir.ui.view">
-        <field name="name">account.move.form.inherit.tier.validation.kmitl</field>
-        <field name="model">account.move</field>
-        <field name="inherit_id" ref="account_move_submitted_state.view_move_form_inherit_submitted_state" />
-        <field name="arch" type="xml">
-
-            <!-- Add id field (invisible) for tier validation -->
-            <xpath expr="//field[@name='id']" position="after">
-                <field name="validation_status" invisible="1" />
-                <field name="need_validation" invisible="1" />
-            </xpath>
-
-
-            <!-- Hide Reset to Draft button when validation is pending -->
-            <xpath expr="//button[@name='action_draft']" position="attributes">
-                <attribute name="attrs">{'invisible': ['|', ('state', '!=', 'submitted'), ('validation_status', 'in', ['validated', 'rejected', 'pending'])]}</attribute>
-            </xpath>
-
-        </field>
-    </record>
-</odoo>
