+++ budget/models/__init__.py
@@ -19,4 +19,3 @@ from . import budget_transfer_reject_wizard
 # Configuration settings
 from . import res_config_settings
 from . import budget_tree
-from . import account_move
